+++ pkg/ccl/sqlccl/backup_cloud_test.go
@@ -1,0 +1,135 @@
+// Copyright 2017 The Cockroach Authors.
+//
+// Licensed under the Cockroach Community Licence (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://github.com/cockroachdb/cockroach/blob/master/pkg/ccl/LICENSE
+
+package sqlccl
+
+import (
+	"fmt"
+	"net/http"
+	"net/url"
+	"os"
+	"testing"
+
+	"github.com/rlmcpherson/s3gof3r"
+
+	"github.com/cockroachdb/cockroach/pkg/ccl/storageccl"
+	"github.com/cockroachdb/cockroach/pkg/sql"
+	"github.com/cockroachdb/cockroach/pkg/storage"
+	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
+	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
+)
+
+// The tests in this file talk to remote APIs which require credentials.
+// To run these tests, you need to supply credentials via env vars (the tests
+// skip themselves if they are not set). Obtain these credentials from the
+// admin consoles of the various cloud providers.
+// customenv.mak (gitignored) may be a useful place to record these.
+// Cockroach Labs Employees: symlink customenv.mk to copy in `production`.
+
+// TestBackupRestoreS3 hits the real S3 and so could occasionally be flaky. It's
+// only run if the AWS_S3_BUCKET environment var is set.
+func TestBackupRestoreS3(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	if !storage.ProposerEvaluatedKVEnabled() {
+		t.Skip("command WriteBatch is not allowed without proposer evaluated KV")
+	}
+	s3Keys, err := s3gof3r.EnvKeys()
+	if err != nil {
+		t.Skipf("No AWS env keys (%v)", err)
+	}
+	bucket := os.Getenv("AWS_S3_BUCKET")
+	if bucket == "" {
+		t.Skip("AWS_S3_BUCKET env var must be set")
+	}
+
+	// TODO(dan): Actually invalidate the descriptor cache and delete this line.
+	defer sql.TestDisableTableLeases()()
+	const numAccounts = 1000
+
+	ctx, _, _, sqlDB, cleanupFn := backupRestoreTestSetup(t, 1, numAccounts)
+	defer cleanupFn()
+	prefix := fmt.Sprintf("TestBackupRestoreS3-%d", timeutil.Now().UnixNano())
+	uri := url.URL{Scheme: "s3", Host: bucket, Path: prefix}
+	values := uri.Query()
+	values.Add(storageccl.S3AccessKeyParam, s3Keys.AccessKey)
+	values.Add(storageccl.S3SecretParam, s3Keys.SecretKey)
+	uri.RawQuery = values.Encode()
+
+	backupAndRestore(ctx, t, sqlDB, uri.String(), numAccounts)
+}
+
+// TestBackupRestoreGoogleCloudStorage hits the real GCS and so could
+// occasionally be flaky. It's only run if the GS_BUCKET environment var is set.
+func TestBackupRestoreGoogleCloudStorage(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	if !storage.ProposerEvaluatedKVEnabled() {
+		t.Skip("command WriteBatch is not allowed without proposer evaluated KV")
+	}
+	bucket := os.Getenv("GS_BUCKET")
+	if bucket == "" {
+		t.Skip("GS_BUCKET env var must be set")
+	}
+
+	// TODO(dan): Actually invalidate the descriptor cache and delete this line.
+	defer sql.TestDisableTableLeases()()
+	const numAccounts = 1000
+
+	// TODO(dt): this prevents leaking an http conn goroutine -- presumably the
+	// conn is held open for reuse until the idle timeout. Ideally we'd test with
+	// conn reuse enabled though, to match expected production behavior.
+	defer func(disableKeepAlives bool) {
+		http.DefaultTransport.(*http.Transport).DisableKeepAlives = disableKeepAlives
+	}(http.DefaultTransport.(*http.Transport).DisableKeepAlives)
+	http.DefaultTransport.(*http.Transport).DisableKeepAlives = true
+
+	ctx, _, _, sqlDB, cleanupFn := backupRestoreTestSetup(t, 1, numAccounts)
+	defer cleanupFn()
+	prefix := fmt.Sprintf("TestBackupRestoreGoogleCloudStorage-%d", timeutil.Now().UnixNano())
+	uri := url.URL{Scheme: "gs", Host: bucket, Path: prefix}
+	backupAndRestore(ctx, t, sqlDB, uri.String(), numAccounts)
+}
+
+// TestBackupRestoreAzure hits the real Azure Blob Storage and so could
+// occasionally be flaky. It's only run if the AZURE_ACCOUNT_NAME and
+// AZURE_ACCOUNT_KEY environment vars are set.
+func TestBackupRestoreAzure(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	if !storage.ProposerEvaluatedKVEnabled() {
+		t.Skip("command WriteBatch is not allowed without proposer evaluated KV")
+	}
+	accountName := os.Getenv("AZURE_ACCOUNT_NAME")
+	accountKey := os.Getenv("AZURE_ACCOUNT_KEY")
+	if accountName == "" || accountKey == "" {
+		t.Skip("AZURE_ACCOUNT_NAME and AZURE_ACCOUNT_KEY env vars must be set")
+	}
+	bucket := os.Getenv("AZURE_CONTAINER")
+	if bucket == "" {
+		t.Skip("AZURE_CONTAINER env var must be set")
+	}
+
+	// TODO(dan): Actually invalidate the descriptor cache and delete this line.
+	defer sql.TestDisableTableLeases()()
+	const numAccounts = 1000
+
+	// TODO(dt): this prevents leaking an http conn goroutine.
+	defer func(disableKeepAlives bool) {
+		http.DefaultTransport.(*http.Transport).DisableKeepAlives = disableKeepAlives
+	}(http.DefaultTransport.(*http.Transport).DisableKeepAlives)
+	http.DefaultTransport.(*http.Transport).DisableKeepAlives = true
+
+	ctx, _, _, sqlDB, cleanupFn := backupRestoreTestSetup(t, 1, numAccounts)
+	defer cleanupFn()
+	prefix := fmt.Sprintf("TestBackupRestoreAzure-%d", timeutil.Now().UnixNano())
+	uri := url.URL{Scheme: "azure", Host: bucket, Path: prefix}
+	values := uri.Query()
+	values.Add(storageccl.AzureAccountNameParam, accountName)
+	values.Add(storageccl.AzureAccountKeyParam, accountKey)
+	uri.RawQuery = values.Encode()
+
+	backupAndRestore(ctx, t, sqlDB, uri.String(), numAccounts)
+}
