+++ pkg/ccl/sqlccl/backup_test.go
@@ -15,19 +15,15 @@ import (
 	"hash/crc32"
 	"io/ioutil"
 	"math/rand"
-	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strconv"
 	"testing"
 
 	"github.com/pkg/errors"
-	"github.com/rlmcpherson/s3gof3r"
 	"golang.org/x/net/context"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
-	"github.com/cockroachdb/cockroach/pkg/ccl/storageccl"
 	"github.com/cockroachdb/cockroach/pkg/internal/client"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
@@ -41,7 +37,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/randutil"
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
-	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 )
 
 const (
@@ -157,105 +152,6 @@ func TestBackupRestoreLocal(t *testing.T) {
 	backupAndRestore(ctx, t, sqlDB, dir, numAccounts)
 }
 
-// TestBackupRestoreS3 hits the real S3 and so could occasionally be flaky. It's
-// only run if the AWS_S3_BUCKET environment var is set.
-func TestBackupRestoreS3(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	s3Keys, err := s3gof3r.EnvKeys()
-	if err != nil {
-		s3Keys, err = s3gof3r.InstanceKeys()
-		if err != nil {
-			t.Skip("No AWS keys instance or env keys")
-		}
-	}
-	bucket := os.Getenv("AWS_S3_BUCKET")
-	if bucket == "" {
-		// CRL uses a bucket `cockroach-backup-tests` that has a 24h TTL policy.
-		t.Skip("AWS_S3_BUCKET env var must be set")
-	}
-
-	// TODO(dan): Actually invalidate the descriptor cache and delete this line.
-	defer sql.TestDisableTableLeases()()
-	const numAccounts = 1000
-
-	ctx, _, _, sqlDB, cleanupFn := backupRestoreTestSetup(t, 1, numAccounts)
-	defer cleanupFn()
-	prefix := fmt.Sprintf("TestBackupRestoreS3-%d", timeutil.Now().UnixNano())
-	uri := url.URL{Scheme: "s3", Host: bucket, Path: prefix}
-	values := uri.Query()
-	values.Add(storageccl.S3AccessKeyParam, s3Keys.AccessKey)
-	values.Add(storageccl.S3SecretParam, s3Keys.SecretKey)
-	uri.RawQuery = values.Encode()
-
-	backupAndRestore(ctx, t, sqlDB, uri.String(), numAccounts)
-}
-
-// TestBackupRestoreGoogleCloudStorage hits the real GCS and so could
-// occasionally be flaky. It's only run if the GS_BUCKET environment var is set.
-func TestBackupRestoreGoogleCloudStorage(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	if !storage.ProposerEvaluatedKVEnabled() {
-		t.Skip("command WriteBatch is not allowed without proposer evaluated KV")
-	}
-
-	bucket := os.Getenv("GS_BUCKET")
-	if bucket == "" {
-		t.Skip("GS_BUCKET env var must be set")
-	}
-
-	// TODO(dan): Actually invalidate the descriptor cache and delete this line.
-	defer sql.TestDisableTableLeases()()
-	const numAccounts = 1000
-
-	// TODO(dt): this prevents leaking an http conn goroutine.
-	http.DefaultTransport.(*http.Transport).DisableKeepAlives = true
-
-	ctx, _, _, sqlDB, cleanupFn := backupRestoreTestSetup(t, 1, numAccounts)
-	defer cleanupFn()
-	prefix := fmt.Sprintf("TestBackupRestoreGoogleCloudStorage-%d", timeutil.Now().UnixNano())
-	uri := url.URL{Scheme: "gs", Host: bucket, Path: prefix}
-	backupAndRestore(ctx, t, sqlDB, uri.String(), numAccounts)
-}
-
-// TestBackupRestoreAzure hits the real Azure Blob Storage and so could
-// occasionally be flaky. It's only run if the AZURE_ACCOUNT_NAME and
-// AZURE_ACCOUNT_KEY environment vars are set.
-func TestBackupRestoreAzure(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	if !storage.ProposerEvaluatedKVEnabled() {
-		t.Skip("command WriteBatch is not allowed without proposer evaluated KV")
-	}
-
-	accountName := os.Getenv("AZURE_ACCOUNT_NAME")
-	accountKey := os.Getenv("AZURE_ACCOUNT_KEY")
-	if accountName == "" || accountKey == "" {
-		t.Skip("AZURE_ACCOUNT_NAME and AZURE_ACCOUNT_KEY env vars must be set")
-	}
-	bucket := os.Getenv("AZURE_CONTAINER")
-	if bucket == "" {
-		t.Skip("AZURE_CONTAINER env var must be set")
-	}
-
-	// TODO(dan): Actually invalidate the descriptor cache and delete this line.
-	defer sql.TestDisableTableLeases()()
-	const numAccounts = 1000
-
-	// TODO(dt): this prevents leaking an http conn goroutine.
-	http.DefaultTransport.(*http.Transport).DisableKeepAlives = true
-
-	ctx, _, _, sqlDB, cleanupFn := backupRestoreTestSetup(t, 1, numAccounts)
-	defer cleanupFn()
-	prefix := fmt.Sprintf("TestBackupRestoreAzure-%d", timeutil.Now().UnixNano())
-	uri := url.URL{Scheme: "azure", Host: bucket, Path: prefix}
-	values := uri.Query()
-	values.Add(storageccl.AzureAccountNameParam, accountName)
-	values.Add(storageccl.AzureAccountKeyParam, accountKey)
-	uri.RawQuery = values.Encode()
-
-	backupAndRestore(ctx, t, sqlDB, uri.String(), numAccounts)
-}
-
 func backupAndRestore(
 	ctx context.Context, t *testing.T, sqlDB *sqlutils.SQLRunner, dest string, numAccounts int64,
 ) {
@@ -266,7 +162,7 @@ func backupAndRestore(
 			&unused, &unused, &unused, &dataSize,
 		)
 		approxDataSize := int64(backupRestoreRowPayloadSize) * numAccounts
-		if max := approxDataSize * 2; dataSize < approxDataSize || dataSize > 2*max {
+		if max := approxDataSize * 2; dataSize < approxDataSize || dataSize > max {
 			t.Errorf("expected data size in [%d,%d] but was %d", approxDataSize, max, dataSize)
 		}
 	}
